pyproject.toml
[tool.poetry]
name = "vrooem-gateway"
version = "0.1.0"
description = "Unified car rental provider gateway — normalizes 12+ supplier APIs into one canonical format"
authors = ["Vrooem <dev@vrooem.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
pydantic = {extras = ["email"], version = "^2.10.0"}
pydantic-settings = "^2.7.0"
httpx = "^0.28.0"
redis = {extras = ["hiredis"], version = "^5.2.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
asyncpg = "^0.30.0"
alembic = "^1.14.0"
pyyaml = "^6.0.2"
jinja2 = "^3.1.4"
python-multipart = "^0.0.18"
lxml = "^5.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
httpx = "^0.28.0"
ruff = "^0.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
